admin routes for sales/vehicle/sellers stats

+++ src/main/java/com/develhope/spring/features/orders/dto/PatchOrderRequest.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PatchOrderRequest {
-    Integer deposit;
+    Long deposit;
     String paymentStatus;
     String orderStatus;
     VehicleEntity vehicleEntity;
