+++ src/main/java/com/develhope/spring/features/vehicle/dto/CreateVehicleRequest.java
@@ -1,9 +1,5 @@
 package com.develhope.spring.features.vehicle.dto;
 
-import com.develhope.spring.features.vehicle.VehicleStatus;
-import com.develhope.spring.features.vehicle.VehicleType;
-import com.develhope.spring.features.vehicle.PropertiesEnum.ShiftType;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -21,7 +17,7 @@ public class CreateVehicleRequest {
     String shiftType;
     Integer yearOfMatriculation;
     String fuelType;
-    Integer price;
+    Long price;
     Integer discount;
     String accessories;
     Boolean used;
